refactor: move PostInspector into a single file
